Criando Participante Service e outras correções

+++ Pactolo/scr/dominio/Pessoa.cs
@@ -10,7 +10,8 @@ namespace Pactolo.scr.dominio {
 	abstract class Pessoa {
 
 		// Id tem que ter set para, ao deletar um objeto, o métodod possa settar seu Id como zero para que se possa saber que aquele registro não existe mais no banco de dados
-		public int Id { get; set; }
+		// Além disso ao salvar uma nova pessoa no banco seu id passa de 0 para seu id no banco
+		public long Id { get; set; }
 
 		protected string nome;
 		public string Nome {
